requirements.txt
pydantic
streamlit
flask
openai
python-dotenv
